+++ src/Audit.EntityFramework.Identity/AuditIdentityDbContext.Net45.cs
@@ -193,6 +193,12 @@ namespace Audit.EntityFramework
         public bool ExcludeTransactionId { get; set; }
 
         /// <summary>
+        /// To indicate if the audit event should be saved before the entity saving operation takes place. 
+        /// Default is false to save the audit event after the entity saving operation completes or fails.
+        /// </summary>
+        public bool EarlySavingAudit { get; set; }
+
+        /// <summary>
         /// Called after the audit scope is created.
         /// Override to specify custom logic.
         /// </summary>