+++ src/Audit.EntityFramework/AuditDbContext.cs
@@ -135,6 +135,12 @@ namespace Audit.EntityFramework
         /// </summary>
         public bool ExcludeTransactionId { get; set; }
 
+        /// <summary>
+        /// To indicate if the audit event should be saved before the entity saving operation takes place. 
+        /// Default is false to save the audit event after the entity saving operation completes or fails.
+        /// </summary>
+        public bool EarlySavingAudit { get; set; }
+
         public DbContext DbContext { get { return this; } }
 #if EF_FULL
         /// <summary>